[FE] Feat : 커스텀 훅 생성 및 스타일 수정
