+++ fe/src/components/molecules/DropdownHeader/style.ts
@@ -7,8 +7,14 @@ export const DropdownButton = styled(Button)`
   color: ${({ theme }) => theme.color.black};
 `;
 
-export const Container = styled.div`
+export interface Prop {
+  theme: any;
+  border?: boolean;
+}
+export const Container = styled.div<Prop>`
   flex: 1;
   position: relative;
   width: calc(100% - 0.6rem - 2px);
+  border-bottom: 1px solid
+    ${({ theme, border }) => (border ? theme.color.subText : 'transparent')};
 `;
