+++ fe/src/components/organisms/FilterBar/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import searchIcon from 'assets/svg/search.svg';
 import Icon from 'components/atoms/Icons';
 import * as S from './style';
+import MainFilterForm from '../MainFilterForm';
 
 export interface Props {}
 
@@ -9,6 +10,7 @@ const FilterBar = ({ ...props }: Props): React.ReactElement => {
   return (
     <S.FilterBar {...props}>
       <div>필터</div>
+      <MainFilterForm />
       <Icon icon={searchIcon} size="sm" />
     </S.FilterBar>
   );
